faq component code refactoring
